% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bbmri.data}
\alias{bbmri.data}
\title{Load BBMRI data from the research drive}
\usage{
bbmri.data(dataset, envir = parent.frame(2))
}
\arguments{
\item{dataset}{The dataset to be loaded.}

\item{envir}{The environment the dataset will be loaded into.}
}
\description{
Load BBMRI data from the research drive
}
\examples{
\dontrun{
# list the available datasets
data(package="BBMRIomics")

# load in a dataset
bbmri.data(metabolomics_RP3RP4_overlap)
}
}
