% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDb_Helpers.R
\name{.getDoc}
\alias{.getDoc}
\title{extract document from couchdb database}
\usage{
.getDoc(id, url, usrpwd, verbose = TRUE)
}
\arguments{
\item{id}{doc._id}

\item{url}{couchdb url}

\item{usrpwd}{username:password}

\item{verbose}{default TRUE}
}
\value{
return list of lists
}
\description{
extract document from couchdb database
}
\details{
R-wrapper around curl to extract document from couchdb database
convert JSON to list of lists
}
\examples{
\dontrun{
doc <- .getDoc("LLS-2749", usrpwd=RP3_MDB_USRPWD, url=RP3_MDB)
names(doc)
str(doc)
}
}
\author{
mvaniterson
}
