% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDb_Helpers.R
\name{.putDoc}
\alias{.putDoc}
\title{put document in couchdb database}
\usage{
.putDoc(doc, url, usrpwd, verbose = TRUE, ...)
}
\arguments{
\item{doc}{list of lists}

\item{url}{couchdb url}

\item{usrpwd}{username:password}

\item{verbose}{default TRUE}

\item{...}{path of database schema}
}
\value{
nothing
}
\description{
put document in couchdb database
}
\details{
converts R list of list to JSON (number of digits = 12)
and validate according to JSON-schema
and if validation succeeds puts document in database
}
\examples{
\dontrun{
doc <- .getDoc("LLS-2749", usrpwd=RP3_MDB_USRPWD, url=RP3_MDB)
##modify or add something
SCHEMA <- file.path(VM_BASE_ANALYSIS, "BBMRIomics/couchdbapp/schema/bios.json")
.validateDoc(doc, SCHEMA=SCHEMA)
##BE REALLY CAREFUL WHAT YOU DO!!!
##.putDoc(doc, usrpwd=RP3_MDB_USRPWD, url=RP3_MDB, SCHEMA=SCHEMA) 
}
}
\author{
mvaniterson
}
