% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDb_Helpers.R
\name{.validateDoc}
\alias{.validateDoc}
\title{validate json documents against schema}
\usage{
.validateDoc(json, SCHEMA)
}
\arguments{
\item{json}{R list representation of json document or json document}

\item{SCHEMA}{path to json schema: bios-schema/bios.json}
}
\value{
error or NULL
}
\description{
validate json document against schema
}
\details{
uses jsonschema CLI Python script from Julian Berman:
https://github.com/Julian/jsonschema
}
\examples{
\dontrun{ 
ids <- getView("getIds", usrpwd=RP3_MDB_USRPWD, url=RP3_MDB)
for(id in ids$ids) {
    print(id)
    doc <- .getDoc(id, usrpwd=RP3_MDB_USRPWD, url=RP3_MDB)
    val <- .validateDoc(doc, SCHEMA=file.path(VM_BASE_ANALYSIS, "BBMRIomics/couchdbapp/schema/bios.json"))
    if(!is.null(val))
        stop(val)
}
}
}
\author{
mvaniterson
}
