% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql_MDb_Helpers.R
\name{getSQLview}
\alias{getSQLview}
\title{get a view from the metadatabase}
\usage{
getSQLview(view, verbose = T, ...)
}
\arguments{
\item{view}{The name of the to be retrieved view.}

\item{...}{Additional arguments to be passed to \link[BBMRIomics]{runQuery}}
}
\value{
A data.frame containing the view.
}
\description{
get a view from the metadatabase
}
\examples{
\dontrun{ 
view <- getSQLview("minimalphenotypes")
}
}
\author{
Davy Cats
}
