% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql_MDb_Helpers.R
\name{getTables}
\alias{getTables}
\title{retrieve table names}
\usage{
getTables(...)
}
\arguments{
\item{...}{Arguments to be passed to \link[BBMRIomics]{runQuery}}
}
\value{
A vector containing the table names.
}
\description{
retrieve table names
}
\examples{
\dontrun{
tables <- getTables()
}
}
\author{
Davy Cats
}
