% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql_MDb_Helpers.R
\name{getViews}
\alias{getViews}
\title{retrieve view names}
\usage{
getViews(...)
}
\arguments{
\item{...}{Arguments to be passed to \link[BBMRIomics]{runQuery}}
}
\value{
A vector containing the view names.
}
\description{
retrieve view names
}
\examples{
\dontrun{
tables <- getViews()
}
}
\author{
Davy Cats
}
