% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{GRID_PROXY}
\alias{GRID_PROXY}
\title{Global variable}
\format{An object of class \code{NULL} of length 0.}
\usage{
GRID_PROXY
}
\description{
Global variable
}
\author{
mvaniterson
}
\keyword{data}

