% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{head}
\alias{head}
\title{head}
\usage{
head(..., p = 6)
}
\arguments{
\item{...}{usual arguments to head}

\item{p}{numebr of columns to show}
}
\value{
selection of rows and columns (in case matrix or data.frame)
}
\description{
head for matrices and data.frames as well
}
\details{
optional p arguments selectes the number of columns to show
}
\author{
mvaniterson
}
