% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql_MDb_Helpers.R
\name{listTables}
\alias{listTables}
\title{retrieve a list of available tables}
\usage{
listTables(...)
}
\arguments{
\item{...}{arguments to be passed to \link[BBMRIomics]{runQuery}}
}
\value{
A vector containing the table names.
}
\description{
retrieve a list of available tables
}
\examples{
\dontrun{
tables <- listTables()
}
}
\author{
Davy Cats
}
