% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql_MDb_Helpers.R
\name{listViews}
\alias{listViews}
\title{retrieve a list of available views}
\usage{
listViews(...)
}
\arguments{
\item{...}{arguments to be passed to \link[BBMRIomics]{runQuery}}
}
\value{
A vector containing the view names.
}
\description{
retrieve a list of available views
}
\examples{
\dontrun{
tables <- listViews()
}
}
\author{
Davy Cats
}
