% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlap_Helpers.R
\name{makeSE}
\alias{makeSE}
\title{construct SummarizedExperiment}
\usage{
makeSE(assaysData, colData, rowData, note = NULL)
}
\arguments{
\item{assaysData}{data matrix}

\item{colData}{phenotype/metadata}

\item{rowData}{feature information}

\item{note}{optional note will be added to exptData}
}
\value{
object of class 'SummarizedExperiment'
}
\description{
construct SummarizedExperiment
}
\details{
construct SummarizedExperiment
the intersection of rows and columns will be used ...
}
\author{
mvaniterson
}

