% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql_MDb_Helpers.R
\name{mdbVersion}
\alias{mdbVersion}
\title{retrieve database version}
\usage{
mdbVersion(db = "rp3_rp4_meta", ...)
}
\arguments{
\item{db}{name of the database to be connected to, defaults to
"rp3_rp4_meta".}

\item{...}{arguments to be passed to \link[BBMRIomics]{runQuery}}
}
\value{
The commit hash for the git repository from which the database 
was built.
}
\description{
retrieve database version
}
\examples{
\dontrun{
v <- mdbVersion()
}
}
\author{
Davy Cats
}
