% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_old.R
\docType{data}
\name{metabolomics_RP3RP4_overlap}
\alias{metabolomics_RP3RP4_overlap}
\title{RP4 metabolites}
\format{
An object of class \code{"SummarizedExperiment"}; see \code{\link[SummarizedExperiment]{SummarizedExperiment}}.
}
\usage{
data(metabolomics_RP3RP4_overlap)
}
\description{
add description
}
\examples{
data(metabolomics_RP3RP4_overlap)
metabolomicData
}
\keyword{datasets}
