% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{methData_Betas_RS_F2_cleaned}
\alias{methData_Betas_RS_F2_cleaned}
\title{DNA methylation data (beta-values) for the RS cohort}
\format{An object of class \code{"SummarizedExperiment"}; see \code{\link[SummarizedExperiment]{SummarizedExperiment}}.}
\usage{
data(methData_Betas_RS_F2_cleaned)
}
\description{
The generation of genome-wide DNA methylation data is described by Bonder et al. (doi:10.1038/ng.3721).
Briefly, 500 ng of genomic DNA was bisulfite modified using the EZ DNA Methylation kit (Zymo Research, Irvine, CA, USA)
and hybridized on Illumina 450K arrays according to the manufacturer’s protocols.
The original idat files were generated by the Illumina iScan BeadChip scanner.
Subsequently, sample quality control was performed using MethylAid.
Probes with a high detection P value (> 0.01), probes with a low bead count (< 3 beads), and probes with a low success rate
(missing in > 95% of the samples) were set to missing. Samples containing an excess of missing probes (> 5%) were excluded
from the analysis. Subsequently, per cohort, imputation was performed to impute the missing values. Functional normalization,
as implemented in the minfi package , was used per cohort.
}
\details{
DNA methylation data and RNA-seq data were generated within the Biobank-based Integrative Omics Studies Consortium (http://wiki.bbmri.nl/wiki/BIOS_start-).
The data comprise six biobanks: Cohort on Diabetes and Atherosclerosis Maastricht (CODAM, n~180)
(doi: 10.1111/j.1365-2362.2010.02418.x), LifeLines (LL, n~700) (doi: 10.1136/bmjopen-2014-006772),
the Leiden Longevity Study (LLS, n~600) (doi: 10.1111/j.1532-5415.2009.02381.x),
the Rotterdam Study (RS, n~600) (doi: 10.1007/s10654-011-9610-5),
the Netherlands Twin Registry (NTR, n ≈ 1800) (doi:  10.1017/thg.2012.140) and
the Prospective ALS Study Netherlands (PAN, n~180) (doi: 10.1136/jnnp.2011.244939).
Sample identity of DNA methylation and gene expression data was confirmed using genotype data.

Data were generated by the Human Genotyping facility (HugeF) of ErasmusMC, the Netherlands (www.glimDNA.org).
}
\examples{
data(methData_Betas_RS_F2_cleaned)
betas
}
\keyword{datasets}

