% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP4_Helpers.R
\name{molgenis.connect}
\alias{molgenis.connect}
\title{connect to a molgenis database using the R-API}
\usage{
molgenis.connect(usrpwd, url, verbose = FALSE)
}
\arguments{
\item{usrpwd}{username:password default reading from .biosrutils}

\item{url}{src source url molgenis database}

\item{verbose}{show additional output}
}
\value{
setup molgenis connection
}
\description{
connect to a molgenis database using the R-API
}
\details{
connect to a molgenis database using the R-API
}
\author{
mvaniterson
}
