% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP4_Helpers.R
\name{molgenis.get.all}
\alias{molgenis.get.all}
\title{get all data}
\usage{
molgenis.get.all(entity, nrows = 1000, verbose = TRUE)
}
\arguments{
\item{entity}{the name of the table to extract data from}

\item{nrows}{number of rows used in the chunking}

\item{verbose}{verbose option default TRUE}
}
\value{
data.frame containing the whole table
}
\description{
extract all data for a particulair entity from a molgenis database
}
\details{
wrapper around molgenis.get
}
\author{
mvaniterson
}

