% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{RP3_BIOBANKS}
\alias{RP3_BIOBANKS}
\title{Global variable}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
RP3_BIOBANKS
}
\description{
Global variable
}
\author{
mvaniterson
}
\keyword{data}
