% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{RP3_MDB_USRPWD}
\alias{RP3_MDB_USRPWD}
\title{Global variable}
\format{An object of class \code{NULL} of length 0.}
\usage{
RP3_MDB_USRPWD
}
\description{
Global variable
}
\author{
mvaniterson
}
\keyword{data}

