% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rp3_rp4_meta}
\alias{rp3_rp4_meta}
\title{RP3 RP4 Metadatabase}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
data(rp3_rp4_meta)
}
\description{
This dataset contains a series of data.frame objects with the metadata for the RP3 and RP4 sample and experiments. 
Including phenotype information. Nightingale measurements are also included.
}
\details{
The dataset contains two types of tables. Tables with raw metadata and tables (or views) with pregenerated subsets/aggregates.
The tables with the raw metadata are dna_sample, gwas, imputation, methylation_450k_freeze, methylation_450k_run, nightingale_quantification, 
nightingale_run, person, relation, rna_freeze, rna_merged_run, rna_run, rna_sample and visit. The remaining objects are views.
}
\examples{
data(rp3_rp4_meta)
visit
}
\keyword{datasets}
