% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sql_MDb_Helpers.R
\name{runQuery}
\alias{runQuery}
\title{send a query to the database}
\usage{
runQuery(query, usrpwd = "guest:guest", url = "localhost", port = 5432,
  db = "rp3_rp4_meta", verbose = T)
}
\arguments{
\item{query}{The query to be send to the database.}

\item{usrpwd}{The username and password concatenated by a colon.
This defaults to "guest:guest".}

\item{url}{The URL through which the database can be accessed. This
defaults to "localhost".}

\item{port}{The port to be used to connect to the database. This defaults
to 5432.}

\item{db}{The name of database to be connected to. This defaults to
"rp3_rp4_meta".}
}
\value{
A data.frame with the query results.
}
\description{
send a query to the database
}
\examples{
\dontrun{
visits <- runQuery("SELECT * FROM visit;", RP3_MDB_USRPWD)
}
}
\author{
Davy Cats
}
