% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{SRM_BASE}
\alias{SRM_BASE}
\title{Global variable}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
SRM_BASE
}
\description{
Global variable
}
\author{
mvaniterson
}
\keyword{data}
