% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDb_Helpers.R
\name{view}
\alias{view}
\title{get view from metadatabase}
\usage{
view(viewname, url = "http://127.0.0.1:5984/mdb_test/",
  converter = "json2csv")
}
\arguments{
\item{viewname}{name of the view}

\item{url}{unique resource location of the metadatabase}

\item{converter}{default json2csv}
}
\value{
metadatabase view converted to a data.frame
}
\description{
get view from metadatabase
}
\details{
extract information from the metadatabase using predefined views
and a list function to convert json to csv.xx
}
\author{
mvaniterson
}

