% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{VM_BASE_DATA}
\alias{VM_BASE_DATA}
\title{Global variable}
\format{An object of class \code{NULL} of length 0.}
\usage{
VM_BASE_DATA
}
\description{
Global variable
}
\author{
mvaniterson
}
\keyword{data}
