% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_Helpers.R
\name{getGenotypes}
\alias{getGenotypes}
\title{extract genotypes from vcf-files}
\usage{
getGenotypes(imputation_id, biobank = c("ALL", "CODAM", "LL", "LLS", "NTR",
  "RS", "PAN"), snps, type = c("GoNL", "HRC", "GoNLv5"), geno = c("GT",
  "DS", "GP"), BASE, ...)
}
\arguments{
\item{imputation_id}{imputation identifier}

\item{biobank}{biobank_id}

\item{snps}{GRanges with snps}

\item{type}{imputation type either "GoNL", "HRC" or "GoNLv5"}

\item{geno}{extract either genotypes, dosages or genotype likelihoods}

\item{BASE}{genotype data location default e.g. VM_BASE_DATA}

\item{...}{optional BPPARAM arguments}
}
\value{
matrix with genotypes
}
\description{
extract genotypes from vcf-files
}
\details{
extract genotypes from vcf-files
given selection of SNPs and samples
}
\author{
mvaniterson
}

