% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDb_Helpers.R
\name{getView}
\alias{getView}
\title{get view from metadatabase}
\usage{
getView(view, url = MDB, usrpwd = USRPWDRP3, selection = "?reduce=false",
  force.keys = FALSE, row.names = FALSE, verbose = TRUE)
}
\arguments{
\item{view}{name of the view}

\item{url}{unique resource location of the metadatabase default MDB}

\item{usrpwd}{username and passwrd concatenated by colon 'usr:pwd'
defaults to USRPWDRP3 if not available the 'anonymous' account will
be used using stored output of the metadatabase}

\item{selection}{some view have a reduce option}

\item{force.keys}{FALSE}

\item{row.names}{use ids as row names in the returned data.frame}

\item{verbose}{default TRUE prints some additional output}
}
\value{
data.frame containing the view
}
\description{
get view from metadatabase
}
\details{
TODO handle reduce options properly
}
\examples{
\dontrun{
##test all views
VIEWS
view <- getView("getIds")
head(view)
colnames(view)
dim(view)
view <- getView("freeze1RNASeq")
head(view)
colnames(view)
dim(view)
view <- getView("allPhenotypes")
head(view)
dim(view)
##For example make a plot of the Age distribution in BIOS
view <- getView("minimalPhenotypes")
library(ggplot2)
view <- view[, c("biobank_id", "sex", "DNA_BloodSampling_Age")]
##convert to numeric
view$DNA_BloodSampling_Age <- as.numeric(view$DNA_BloodSampling_Age)
str(as.numeric(view$DNA_BloodSampling_Age))
g <- ggplot(view, aes(x = DNA_BloodSampling_Age))
g <- g + geom_histogram()
g <- g + facet_grid(. ~ biobank_id)
g
g <- g + facet_grid(sex ~ biobank_id)
g
}
}
\author{
mvaniterson
}
