% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SRM2VM}
\alias{SRM2VM}
\title{copy file from SRM2VM}
\usage{
SRM2VM(srm.path, files, vm.path, proxy)
}
\arguments{
\item{srm.path}{srm path}

\item{files}{file or vector of files}

\item{vm.path}{vm path}

\item{proxy}{a valid proxy from example PROXY from your global
envirnoment}
}
\value{
curl stdout message
}
\description{
Utility function to copy files from SRM to VM
}
\details{
details follow
}
\author{
mvaniterson
}
