% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SRM2VM}
\alias{SRM2VM}
\title{copy file from SRM2VM}
\usage{
SRM2VM(srm.path, files, vm.path, proxy)
}
\arguments{
\item{srm.path}{srm path}

\item{files}{file or vector of files}

\item{vm.path}{vm path}

\item{proxy}{a valid proxy from example PROXY from your global
envirnoment}
}
\value{
curl stdout message
}
\description{
Utility function to copy files from SRM to VM
}
\details{
details follow
}
\examples{
\dontrun{
##get fastq
fastq <- getView("getFastq", usrpwd=RP3_MDB_USRPWD, url=RP3_MDB)
file <- basename(fastq$R1[1])
srm.path <- file.path(dirname(fastq$R1[1])) 
srm.path <- gsub("srm.*nl", SRM_BASE, srm.path) ##for curl access with need slightly different url 
srm.path
files
SRM2VM(srm.path, files, vm.path="~", proxy=GRID_PROXY)
##get bam
bam <- getView("getBAM", usrpwd=RP3_MDB_USRPWD, url=RP3_RDB)
bam$path[1]
file <- basename(bam$path[1])
srm.path <- file.path(dirname(bam$path[1])) 
srm.path <- gsub("srm.*nl", SRM_BASE, srm.path) ##for curl access with need slightly different url 
srm.path
file
SRM2VM(srm.path, file, vm.path="~", proxy=GRID_PROXY)
}
}
\author{
mvaniterson
}
