% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDb_Helpers.R
\name{view}
\alias{view}
\title{get view from metadatabase}
\usage{
view(viewname,
  db = "https://metadatabase.bbmrirp3-lumc.surf-hosted.nl:6984/bios-test/",
  converter = "json2csv", usrpwd = NULL)
}
\arguments{
\item{viewname}{name of the view}

\item{db}{metadatabase url}

\item{converter}{default json2csv}

\item{usrpwd}{metadata usrpwd}
}
\value{
metadatabase view converted to a data.frame
}
\description{
get view from metadatabase
}
\details{
extract information from the metadatabase using predefined views
and a list function to convert json to csv.xx
}
\author{
mvaniterson
}
